rename: dialog -> capsulepreview, /ui/home -> /ui
