+++ frontend/ARchive/app/src/main/res/layout/fragment_capsule_preview_dialog.xml
@@ -14,7 +14,7 @@
 
         <variable
             name="vm"
-            type="com.droidblossom.archive.presentation.ui.home.dialog.CapsulePreviewDialogViewModel" />
+            type="com.droidblossom.archive.presentation.ui.capsulepreview.CapsulePreviewDialogViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
